feature/95412 - feature/95410

See merge request amcom-pedagogico/sme-acessos!28

+++ SME.Acessos.Infra.Dominio/Acessos/Repositorios/IRepositorioSistemaRecuperacaoSenha.cs
@@ -5,5 +5,5 @@ namespace SME.Acessos.Infra.Dominio.Acessos.Repositorios;
 
 public interface IRepositorioSistemaRecuperacaoSenha
 {
-    Task<SistemaRecuperacaoSenha> ObterSistema(int sistemaId);
+    Task<SistemaRecuperacaoSenha> ObterSistema(long sistemaId);
 }
